+++ src/frontend/apps/impress/src/features/docs/doc-export/blocks-mapping/uploadLoaderODT.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { DocsExporterODT } from '../types';
+import { type DocsExporterODT } from '../types';
 
 export const blockMappingUploadLoaderODT: DocsExporterODT['mappings']['blockMapping']['uploadLoader'] =
   (block) => {