+++ CHANGELOG.md
@@ -75,6 +75,7 @@ and this project adheres to
   `/external_api/{version}/jwks`
 - 🔧(collaboration) adapt docker stack for development purpose
 - 🔧(helm) run a valkey for the backend and one for yhub in dev and feature
+- ✨(frontend) turn pasted doc links into interlinks #2713
 
 ### Fixed